Refactor: Mess clean up
